Adding checks against package.json

+++ git-template-dir/hooks/pre-release.sh
@@ -1,1 +1,3 @@
-node -e "f = './package.json'; p = require(f); p.version = process.argv[1]; require('fs').writeFileSync(f, JSON.stringify(p, null, 2));" $2
+if [ test -f package.json ]; then
+  node -e "f = './package.json'; p = require(f); p.version = process.argv[1]; require('fs').writeFileSync(f, JSON.stringify(p, null, 2));" $2
+fi
